+++ src/test/resources/fixtures/consented-ccd-request-with-solicitor-agreed-to-emails.json
@@ -125,6 +125,7 @@
       "divorceDecreeAbsoluteDate": "2010-01-01",
       "orderForChildrenQuestion1": "Yes",
       "solicitorAgreeToReceiveEmails": "Yes",
+      "RespSolNotificationsEmailConsent": "Yes",
       "orderDirection": "test",
       "orderDirectionOpt1": {
         "document_url": "http://file1",
